# StrnadiAPI-openapi.yaml
openapi: 3.0.1
info:
  title: Strnadi API
  description: |
    HTTP API used by the Strnadi applications to manage user accounts, bird recordings,
    articles and other shared resources. The specification documents the behavior of the
    existing ASP.NET controllers so it can be imported into Swagger UI or other OpenAPI tooling.
  version: 1.0.0
servers:
  - url: https://api.strnadi.cz
    description: PROD API
  - url: https://devapi.strnadi.cz
    description: DEV API
security:
  - bearerAuth: []
...
paths:
...
  /utils/fix-same-dates:
    get:
      tags: [Utils]
      summary: Fix duplicate recording part timestamps
      description: Re-runs the maintenance routine that normalizes duplicate timestamps for recording parts.
      responses:
        '200':
          description: The maintenance routine completed.
        '400':
          description: Missing bearer token.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid JWT supplied or the caller is not an administrator.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /utils/normalize-existing-audios:
    get:
      tags: [Utils]
      summary: Normalize stored recording audio files
      description: Iterates through all persisted recording parts and normalizes their audio levels.
      responses:
        '200':
          description: Audio normalization was triggered successfully.
        '400':
          description: Missing bearer token.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid JWT supplied or the caller is not an administrator.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /utils/analyze-parts:
    get:
      tags: [Utils]
      summary: Analyze stored recording parts
      description: Logs technical metadata for each stored recording part such as duration and format.
      responses:
        '200':
          description: Recording part analysis was triggered successfully.
        '400':
          description: Missing bearer token.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid JWT supplied or the caller is not an administrator.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: string
      description: Human readable message describing the error.
...